django/requirements/base.txt
setuptools>=0.8
Django==1.6
factory-boy==2.2.1
pytest-cov==2.2.0
pytest-sugar==0.5.1
pytest-django==2.9.1
pytest==2.8.7
Sphinx==1.2b3
mock==1.0.1
South==1.0.2
django-mptt==0.6.1
django-allauth==0.18.0
django-bootstrap-form==3.1
pillow==2.7
easy-thumbnails==2.2
Markdown==2.6.1
django-taggit==0.12.3
python-dateutil
django-ses==0.7.0
boto==2.38.0
slackclient==1.0.0
